requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
img2pdf>=0.5.1
pillow>=10.3.0
python-dotenv>=1.0.1
pydantic>=2.7.0
playwright
beautifulsoup4
jinja2
